fix: refactor internal/cache pkg to pkg/cache
