+++ pkg/cache/cache_opts_test.go
