+++ pkg/cache/image_cache.go
